Moved runSubproblem
